Audio: mux: move mux header file to mux module folder
Move mux header file to mux module folder.
