+++ src/audio/mux/mux_generic.c
@@ -10,12 +10,13 @@
 #include <sof/audio/buffer.h>
 #include <sof/audio/component.h>
 #include <sof/audio/format.h>
-#include <sof/audio/mux.h>
 #include <rtos/bit.h>
 #include <sof/common.h>
 #include <ipc/stream.h>
 #include <stddef.h>
 #include <stdint.h>
+
+#include "mux.h"
 
 LOG_MODULE_DECLARE(muxdemux, CONFIG_SOF_LOG_LEVEL);
